spec 18: the declaration layer — schema registry, the C6 package contract (77), the C3 driver layout check (78)
Three cohesive pieces of the spec-18 declaration layer:

1. docs/spec/schemas/ — the versioned YAML schema registry, the single
   owner of every grammar in spec 18 §2.2: artifact-info, runtime-manifest,
   layout, payload-manifest, package-manifest, tpkg-registry, store-config,
   versions (the ruby repo's). Each carries schema:/schema_version:/
   schema_minor: + the §3 evolution metadata (major/minor semantics,
   deprecation window, the critical-flag rule). Existing grammars are
   formalized as-is (schema_version 1) — declaration, not redesign.

2. The L2 package contract block (C6): the type-2 block gains
   contract_era / pressed_by / reader_era as YAML keys (additive — the
   composition model is unchanged, so pre-contract readers tolerate them).
   set_package_manifest is the single emission point: every press declares
   era 2. The typed tpkg::ContractError separates "no block" (era 1:
   "pre-era package — re-press with tebako >= 0.16.1 or run it with tebako
   v1") from era mismatch ("package from a newer tebako (era N) — upgrade
   your tebako (speaks era M)"; reader_era too new) — every variant maps
   to exit 77, wired into tebako-pkg validate / info --verify. The
   critical flag (high bit of the ext-block type word, wire-compatible):
   an unknown critical block refuses (§3.7/S10); non-critical unknowns
   still skip (invariant 7).

3. The driver layout check (C3/S17-S19): post-mount, before any
   interpreter handoff, the driver reads /lib/tebako/layout.yaml from the
   mounted env image through the tfs layer and verifies it against the
   exe's expectations — absent → era-1 refusal naming the image; newer
   schema/era/image_layout → upgrade refusal; mount_root != the exe's
   root → refusal with both values. All exit 78, wired through ffi +
   driver::boot; the 65-76 allocation untouched.

tebako-pkg's verify-stage fixtures now press era-2 packages (the strict
gate's contract stage must pass to reach the 65/70/71/72 checks); no
golden byte changes (goldens carry no type-2 blocks). tfs gets four
platform-lint allowances for the local 1.94 clippy leg (i64::from(off_t)
is a real widening on mingw — allow, not removal).

+++ crates/tfs/src/exec_closure.rs
@@ -278,7 +278,7 @@ fn parse_elf(bytes: &[u8]) -> Option<ImageDeps> {
             )
         } else {
             (
-                u32e(bytes.get(base..base + 4)?)? as u32 as i32 as i64,
+                u32e(bytes.get(base..base + 4)?)? as i32 as i64,
                 u32e(bytes.get(base + 4..base + 8)?)? as u64,
             )
         };
